% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ghana_education.R
\name{get_ghana_education}
\alias{get_ghana_education}
\title{Get Ghana Education Data from World Bank}
\usage{
get_ghana_education(start = 2000, end = 2023)
}
\arguments{
\item{start}{Start year (default 2000)}

\item{end}{End year (default 2023)}
}
\value{
A cleaned tibble of Ghana education indicators by year
}
\description{
Get Ghana Education Data from World Bank
}
\examples{
data <- get_ghana_education(start = 2000, end = 2023)
}
