% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_west_africa_education.R
\name{get_west_africa_education}
\alias{get_west_africa_education}
\title{Get West Africa Education Data from World Bank}
\usage{
get_west_africa_education(start = 2000, end = 2023)
}
\arguments{
\item{start}{Start year (default 2000)}

\item{end}{End year (default 2023)}
}
\value{
A cleaned tibble of education indicators for West African countries
}
\description{
Get West Africa Education Data from World Bank
}
\examples{
data <- get_west_africa_education(start = 2000, end = 2023)
}
