% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_enrollment.R
\name{plot_enrollment}
\alias{plot_enrollment}
\title{Plot Primary vs Secondary Enrollment Trends}
\usage{
plot_enrollment(data)
}
\arguments{
\item{data}{A tibble from \code{get_ghana_education()}}
}
\value{
A ggplot2 object
}
\description{
Plot Primary vs Secondary Enrollment Trends
}
\examples{
data <- get_ghana_education()
plot_enrollment(data)
}
