% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gender_gap.R
\name{plot_gender_gap}
\alias{plot_gender_gap}
\title{Plot Gender Gap in Primary Enrollment}
\usage{
plot_gender_gap(data)
}
\arguments{
\item{data}{A tibble from \code{get_ghana_education()}}
}
\value{
A ggplot2 object
}
\description{
Plot Gender Gap in Primary Enrollment
}
\examples{
data <- get_ghana_education()
plot_gender_gap(data)
}
