% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_out_of_school.R
\name{plot_out_of_school}
\alias{plot_out_of_school}
\title{Plot Out-of-School Children Rate}
\usage{
plot_out_of_school(data)
}
\arguments{
\item{data}{A tibble from \code{get_ghana_education()}}
}
\value{
A ggplot2 object
}
\description{
Plot Out-of-School Children Rate
}
\examples{
data <- get_ghana_education()
plot_out_of_school(data)
}
