% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pupil_teacher.R
\name{plot_pupil_teacher}
\alias{plot_pupil_teacher}
\title{Plot Pupil-Teacher Ratio}
\usage{
plot_pupil_teacher(data)
}
\arguments{
\item{data}{A tibble from \code{get_ghana_education()}}
}
\value{
A ggplot2 object
}
\description{
Plot Pupil-Teacher Ratio
}
\examples{
data <- get_ghana_education()
plot_pupil_teacher(data)
}
