% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_regional_comparison.R
\name{plot_regional_comparison}
\alias{plot_regional_comparison}
\title{Plot Regional Comparison of Primary Enrollment}
\usage{
plot_regional_comparison(data)
}
\arguments{
\item{data}{A tibble from \code{get_west_africa_education()}}
}
\value{
A ggplot2 object
}
\description{
Plot Regional Comparison of Primary Enrollment
}
\examples{
data <- get_west_africa_education()
plot_regional_comparison(data)
}
